README.md
I wrote this script in order to save the trouble of writing a series of commands while using the nmap tool. This script is very powerful in scanning websites and IP addresses for vulnerabilities, and if there are vulnerabilities that can be exploited, the exploitation will be done as well. I would very much like to share my simple work with you.